requirements.txt
numpy==2.0.2
opencv-python==4.12.0.88
pyrealsense2-macosx==2.54.2
